requirements.txt
# Intentionally vulnerable AI/ML dependencies for AISBOM scanning demo
# DO NOT USE IN PRODUCTION - These versions have known CVEs

# Core AI/ML frameworks with known vulnerabilities
tensorflow==2.10.0          # CVE-2023-25801, CVE-2023-25660, CVE-2023-25671
torch==1.12.0               # CVE-2022-45907 (arbitrary code execution)
transformers==4.25.1        # CVE-2023-28009 (unsafe deserialization)
langchain==0.0.27           # CVE-2023-36258 (arbitrary code execution via prompt injection)

# Supporting ML libraries with vulnerabilities
numpy==1.21.0               # CVE-2021-41496 (buffer overflow)
scikit-learn==1.0.0         # Older version for SBOM detection
scipy==1.7.0                # CVE-2023-25399
pillow==8.3.0               # CVE-2022-22815, CVE-2022-22816 (multiple image processing vulnerabilities)
opencv-python==4.5.3.56     # Multiple CVEs in older versions

# Model runtime and serving
onnxruntime==1.12.0         # CVE-2023-32708
protobuf==3.19.0            # CVE-2022-1941 (parsing vulnerabilities)

# AI application frameworks
gradio==3.0                 # Multiple CVEs in UI components
openai==0.27.0              # Older version
huggingface-hub==0.10.0     # Older version

# Web framework and utilities
flask==2.0.0                # CVE-2023-30861 (Werkzeug security)
requests==2.25.0            # CVE-2023-32681 (proxy credential exposure)
jinja2==3.0.0               # CVE-2024-22195
pyyaml==5.3                 # CVE-2020-14343 (arbitrary code execution)
pickle5==0.0.11             # Demonstrating insecure deserialization

# Additional vulnerable dependencies
cryptography==3.3.2         # CVE-2023-23931
urllib3==1.26.5             # CVE-2023-45803
certifi==2021.5.30          # Outdated certificate bundle
setuptools==57.0.0          # CVE-2022-40897

# ML ops and data processing
pandas==1.3.0               # CVE-2023-xxxxx
datasets==1.18.0            # Older HuggingFace datasets
tokenizers==0.13.0          # Older version
